refactor(pdf): extract region column layout
